# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API docs for Auto Trade",
        "title": "TOC MACHINE TRADING",
        "contact": {},
        "version": "2.5.0"
    },
    "paths": {
...
        "/v1/basic/config": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "getAllConfig",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Config"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/basic/stock": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basic"
                ],
                "summary": "getAllRepoStock",
                "parameters": [
                    {
                        "type": "string",
                        "description": "num",
                        "name": "num",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.stockDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/basic/usage/shioaji": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "getShioajiUsage",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.ShioajiUsage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/v1/history/day-kbar/{stock}/{start_date}/{interval}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "history"
                ],
                "summary": "getKbarData",
                "parameters": [
                    {
                        "type": "string",
                        "description": "stock",
                        "name": "stock",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "start_date",
                        "name": "start_date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "interval",
                        "name": "interval",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.StockHistoryKbar"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.AnalyzeStock": {
            "type": "object",
            "properties": {
                "allInOutRatio": {
                    "type": "number"
                },
                "allOutInRatio": {
                    "type": "number"
                },
                "closeChangeRatioHigh": {
                    "type": "number"
                },
                "closeChangeRatioLow": {
                    "type": "number"
                },
                "maperiod": {
                    "type": "integer"
                },
                "maxHoldTime": {
                    "type": "number"
                },
                "rsiminCount": {
                    "type": "integer"
                },
                "tickAnalyzePeriod": {
                    "type": "number"
                },
                "volumePRLimit": {
                    "type": "number"
                }
            }
        },
        "config.Config": {
            "type": "object",
            "properties": {
                "analyzeStock": {
                    "$ref": "#/definitions/config.AnalyzeStock"
                },
                "database": {
                    "$ref": "#/definitions/config.Database"
                },
                "fugle": {
                    "$ref": "#/definitions/config.Fugle"
                },
                "history": {
                    "$ref": "#/definitions/config.History"
                },
                "manualTrade": {
                    "type": "boolean"
                },
                "quota": {
                    "$ref": "#/definitions/config.Quota"
                },
                "rabbitMQ": {
                    "$ref": "#/definitions/config.RabbitMQ"
                },
                "server": {
                    "$ref": "#/definitions/config.Server"
                },
                "simulation": {
                    "type": "boolean"
                },
                "sinopac": {
                    "$ref": "#/definitions/config.Sinopac"
                },
                "smtp": {
                    "$ref": "#/definitions/config.SMTP"
                },
                "targetStock": {
                    "$ref": "#/definitions/config.TargetStock"
                },
                "tradeFuture": {
                    "$ref": "#/definitions/config.TradeFuture"
                },
                "tradeStock": {
                    "$ref": "#/definitions/config.TradeStock"
                }
            }
        },
        "config.Database": {
            "type": "object",
            "properties": {
                "dbname": {
                    "type": "string"
                },
                "poolMax": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "config.Fugle": {
            "type": "object",
            "properties": {
                "poolMax": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "config.History": {
            "type": "object",
            "properties": {
                "historyClosePeriod": {
                    "type": "integer"
                },
                "historyKbarPeriod": {
                    "type": "integer"
                },
                "historyTickPeriod": {
                    "type": "integer"
                }
            }
        },
        "config.PriceLimit": {
            "type": "object",
            "properties": {
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                }
            }
        },
        "config.Quota": {
            "type": "object",
            "properties": {
                "futureTradeFee": {
                    "type": "integer"
                },
                "stockFeeDiscount": {
                    "type": "number"
                },
                "stockTradeQuota": {
                    "type": "integer"
                }
            }
        },
        "config.RabbitMQ": {
            "type": "object",
            "properties": {
                "attempts": {
                    "type": "integer"
                },
                "exchange": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "waitTime": {
                    "type": "integer"
                }
            }
        },
        "config.SMTP": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "config.Server": {
            "type": "object",
            "properties": {
                "disableSwaggerHTTPHandler": {
                    "type": "string"
                },
                "http": {
                    "type": "string"
                }
            }
        },
        "config.Sinopac": {
            "type": "object",
            "properties": {
                "poolMax": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "config.TargetStock": {
            "type": "object",
            "properties": {
                "blackCategory": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "blackStock": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "limitVolume": {
                    "type": "integer"
                },
                "priceLimit": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/config.PriceLimit"
                    }
                },
                "realTimeRank": {
                    "type": "integer"
                }
            }
        },
        "config.TradeFuture": {
            "type": "object",
            "properties": {
                "allowTrade": {
                    "type": "boolean"
                },
                "buySellWaitTime": {
                    "type": "integer"
                },
                "inOutRatio": {
                    "type": "number"
                },
                "maxHoldTime": {
                    "type": "integer"
                },
                "outInRatio": {
                    "type": "number"
                },
                "quantity": {
                    "type": "integer"
                },
                "rateChangeRatio": {
                    "type": "number"
                },
                "rateLimit": {
                    "type": "number"
                },
                "subscribe": {
                    "type": "boolean"
                },
                "targetBalanceHigh": {
                    "type": "number"
                },
                "targetBalanceLow": {
                    "type": "number"
                },
                "tickInterval": {
                    "type": "integer"
                },
                "tradeOutWaitTimes": {
                    "type": "integer"
                },
                "tradeTimeRange": {
                    "$ref": "#/definitions/config.TradeTimeRange"
                }
            }
        },
        "config.TradeStock": {
            "type": "object",
            "properties": {
                "allowTrade": {
                    "type": "boolean"
                },
                "cancelWaitTime": {
                    "type": "integer"
                },
                "holdTimeFromOpen": {
                    "type": "number"
                },
                "odd": {
                    "type": "boolean"
                },
                "subscribe": {
                    "type": "boolean"
                },
                "totalOpenTime": {
                    "type": "number"
                },
                "tradeInEndTime": {
                    "type": "number"
                },
                "tradeInWaitTime": {
                    "type": "integer"
                },
                "tradeOutWaitTime": {
                    "type": "integer"
                }
            }
        },
        "config.TradeTimeRange": {
            "type": "object",
            "properties": {
                "firstPartDuration": {
                    "type": "integer"
                },
                "secondPartDuration": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ShioajiUsage": {
            "type": "object",
            "properties": {
                "connections": {
                    "type": "integer"
                },
                "traffic_usage": {
                    "type": "number"
                },
                "traffic_usage_percents": {
                    "type": "number"
                }
            }
        },
        "entity.Stock": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "code": {
                    "type": "string"
                },
                "day_trade": {
                    "type": "boolean"
                },
                "delivery_date": {
                    "type": "string"
                },
                "delivery_month": {
                    "type": "string"
                },
                "exchange": {
                    "type": "string"
                },
                "last_close": {
                    "type": "number"
                },
                "limit_down": {
                    "type": "number"
                },
                "limit_up": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                },
                "reference": {
                    "type": "number"
                },
                "symbol": {
                    "type": "string"
                },
                "underlying_kind": {
                    "type": "string"
                },
                "unit": {
                    "type": "integer"
                },
                "update_date": {
                    "type": "string"
                }
            }
        },
        "entity.StockHistoryKbar": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "kbar_time": {
                    "type": "string"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "stock": {
                    "$ref": "#/definitions/entity.Stock"
                },
                "stock_num": {
                    "type": "string"
                },
                "volume": {
                    "type": "integer"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "response": {
                    "type": "string"
                }
            }
        },
...
        "v1.stockDetailResponse": {
            "type": "object",
            "properties": {
                "stock_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Stock"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}